frontend/server/requirements.txt
pandas
pymongo[srv]
certifi
flask
flask-restx
datetime
python-dotenv